+++ src/main/java/com/bookverser/BookVerse/dto/SearchBooksRequestDTO.java
@@ -3,6 +3,8 @@ package com.bookverser.BookVerse.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
 /**
@@ -17,8 +19,8 @@ public class SearchBooksRequestDTO {
     private String keyword;
 
     @PositiveOrZero(message = "Minimum price cannot be negative")
-    private Double minPrice;
+    private BigDecimal minPrice;
 
     @PositiveOrZero(message = "Maximum price cannot be negative")
-    private Double maxPrice;
+    private BigDecimal maxPrice;
 }
